Added what I believe to be the intended functionality - the system uses uint8/16 to force the byte(s) to fit the desired output, then uses double to generate a physical output in line with what it appears we should be seeing.'

+++ Curs.m
@@ -1,3 +1,3 @@
 function [ Cur ] = Curs(byte)
 
-Cur = byte*0.064; % Result is in amps
+Cur = double(uint8(byte))*0.064; % Result is in amps
