+++ VoltsBatt.m
@@ -1,3 +1,3 @@
 function [ Volt ] = VoltsBatt(byte)
 
-Volt = (16/255)*byte; % Volts
+Volt = (16/255)*double(uint8(byte)); % Volts
